Issue-3499: Move utils module to compose/common package
Partially implements #3499
