+++ compose/cli/log_printer.py
@@ -11,9 +11,9 @@ from six.moves.queue import Empty
 from six.moves.queue import Queue
 
 from . import colors
-from compose import utils
+from compose.common import utils
 from compose.common.signals import ShutdownException
-from compose.utils import split_buffer
+from compose.common.utils import split_buffer
 
 
 class LogPresenter(object):
