+++ compose/common/progress_stream.py
@@ -1,7 +1,7 @@
 from __future__ import absolute_import
 from __future__ import unicode_literals
 
-from compose import utils
+from compose.common import utils
 
 
 class StreamOutputError(Exception):
