+++ tests/unit/utils_test.py
@@ -2,7 +2,7 @@
 from __future__ import absolute_import
 from __future__ import unicode_literals
 
-from compose import utils
+from compose.common import utils
 
 
 class TestJsonSplitter(object):
